// src/api/generated/ActivityView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ActivityItem } from "./ActivityItem";

export type ActivityView = { active: Array<ActivityItem>, queued: Array<ActivityItem>, recent: Array<ActivityItem>, generatedAt: string, };